==> Experiments/0.LoadPDF_data.ipynb <==
# Import libraries


!pip install -U langchain-community
!pip install pypdf
!pip install rapidocr-onnxruntime # If we want to also extract images from the pdf, but the process is much slower

# Try using also pdfplumber
!pip install pdfplumber

!pip install tabula-py

# Load PDF data with langchain


from langchain_community.document_loaders import PyPDFLoader

loader = PyPDFLoader(
    file_path = "../Data/prova_immagini.pdf",
    # password = "my-pasword",
    extract_images = True,
    # headers = None
    # extraction_mode = "plain",
    # extraction_kwargs = None,
)

docs = []
pages = loader.load_and_split()

# Extract only page_content from the pages
for page in pages:
    docs.append(page.page_content)
    # docs.append(page.metadata) - In case we decide we need the metadata

pages[2]

pages[1]

len(docs)

docs[11]

# Load data with pdfplumber

import pdfplumber
filename = "../Data/testdoc.pdf"
pdf = pdfplumber.open(filename)
text = pdf.pages[6].extract_text(x_tolerance=1)
print(text)

# Extract the text from all pages
texts = [page.extract_text(x_tolerance=1) for page in pdf.pages]

# Can concatanate all the text together in a single string
text = "  \n\n".join(texts)
print(text)

# Extract images

from pypdf import PdfReader

reader = PdfReader("../Data/prova_immagini.pdf")
page = reader.pages[0]

for count, image_file_object in enumerate(page.images):
  with open(str(count) + image_file_object.name, "wb") as fp:
    fp.write(image_file_object.data)

for page in reader.pages:
  for count, image_file_object in enumerate(page.images):
    print(image_file_object)
    with open(str(count) + str(page.page_number) + image_file_object.name , "wb") as fp:

      fp.write(image_file_object.data)

# Extracting tables from pdf

filename = "../Data/prova_table.pdf"

with pdfplumber.open(filename) as pdf:
    page = pdf.pages[14]
    table = page.extract_table()
    print(table)

with pdfplumber.open(filename) as pdf:
    for page in pdf.pages:
        table = page.extract_table()
        print(table)
